Update command-line args
The program now accepts either a single path as argument, the option -f
with a path argument, or -x with a Brainfuck string as argument.
The path argument points to a file from where the Brainfuck program will
be read.

To parse the arguments and retrieve the Brainfuck program, a new
function is declared, read_program, that takes as parameters main's
arguments along with a pointer to the program string and to a flag used
to indicate whether the program string needs to be freed.
A new error code is added, and is returned from the new function (and
then from the program) when memory for the program cannot be allocated.