Reorder, assign and document error codes
The error codes are arranged in sections of related error causes.
Generic codes are added, these are not actually returned by the program
but serve as bases for the error-specific codes.
The codes are given explicit values, where the high order bits are used
to encode to which the section the error code belongs and the low order
bits are used to give the error code its unique value.

The codes are also given a short explanation comment.