Add comments explaining how to implement circular tape
The comments are added in the blocks corresponding to the < and >
Brainfuck commands in the main switch in function run.
They state that merely wrapping back the tape pointer to 0 or
(TAPE_SIZE - 1) instead of returning one of the tape overflow error
codes is enough to implement such a circular tape.