Rework the loop bounds arrays
The arrays no longer represent the opening and closing bracket pointers
in parallel manner, but rather each loop bound (opening and closing
alike) and the pointer difference to the matching counterpart bound (in
other words, the jump length).
The code is much more complex than before, but now it works correctly.