Implement -h, --help command-line option
A new constant is added to the error codes enum to indicate that usage
output was requested by the caller. It is returned from read_program
when the command-line arguments consist of only -h or --help.
It is not returned from main however, but the success code instead, as
it is not a failure.